improve unit and integration tests
